refactor: extract the demographic reading behind a per-pre-test config [REPORT-81]
Move the answer query, prompt matching, choice resolution and category
mapping into demographics.ts, parameterised by a PreTestConfig (the
prompts and choice labels) and by the set of dimensions to read. The
wording moves to pre-tests.ts and the 24-row table to strata-tables.ts,
leaving random-assignment.ts holding orchestration only, 450 lines down
to 146.

Selection is structural: the step references its own wording object,
because the pipeline already selected the step and the step already knows
its pre-test. Nothing is keyed by pilot and nothing is resolved at run
time.

readDemographics separates three failure kinds by whether retrying can
help: incomplete (the student has not answered), unmappable (an authoring
fault, permanent) and failed (transport, transient). Spring maps the last
two to its one existing message, so its behaviour is unchanged.

One deliberate, behaviour-visible difference: the client Firestore
instance is now released as soon as the answers have been read rather
than after the enrollment call, which is strictly narrower. The
enrollment path never used it.

NO TEST FILE IS EDITED, and that is the evidence the extraction preserves
behaviour: 9 suites, 231 tests, all green unmodified.
